+++ App.js
@@ -1,17 +1,23 @@
 import { createStackNavigator, createSwitchNavigator, createBottomTabNavigator, getActiveChildNavigationOptions } from 'react-navigation';
 
-import Reading from './screens/reading';
-import Library from './screens/library';
-import Words from './screens/words';
+import Reading from './screens/readingView';
+import Library from './screens/libraryList';
+import Bookshelf from './screens/readingList';
+import Words from './screens/wordMagician';
 import FlashCard from './screens/flashCard';
 
 const TabNavigator = createBottomTabNavigator(
   {
-    Library: {
-      screen: Library,
+    Bookshelf: {
+      screen: Bookshelf,
       navigationOptions: ({ navigation }) => ({
-        title: '📚'
+        title: '📖'
       }),},
+    Library: {
+    screen: Library,
+    navigationOptions: ({ navigation }) => ({
+      title: '📚'
+    }),},
     Words: {
       screen: Words,
       navigationOptions: ({ navigation }) => ({
